requirements.txt
Click
docker
GitPython
